requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
